refactor: reorganize feedback components into feature directory
- Move feedback components from src/app/components/feedback/ to src/app/feedback/components/
- Move FeedbackProvider from src/app/components/providers/ to src/app/feedback/
- Move FeedbackWidget to src/app/feedback/ for logical grouping
- Update import paths in providers.tsx and layout.tsx
- Remove empty directories
- Improve code organization following Next.js App Router feature-based structure
